requirements.txt
rich
colour
tqdm
requests
pytz
spotipy
json
datetime
pillow
tweepy
python-telegram-bot